growing-oo-software-guided-by-tests/python/requirements.txt
nbxmpp
xmpppy
pytest-qt
pyhamcrest
